finish search page refactory the project structure
