+++ src/App.tsx
@@ -1,6 +1,6 @@
 import "./App.css";
 import "react-h5-audio-player/lib/styles.css";
-import { AppRoutes } from "./routes";
+import { AppRoutes } from "./layout/routes/routes";
 function App() {
   return (
     <div>
