+++ src/layout/Layout.tsx
@@ -1,7 +1,7 @@
+import { Footer } from "@/components/Footer";
+import { Header } from "@/components/Header";
+import { SideBar } from "@/components/SideBar";
 import { Outlet } from "react-router-dom";
-import { Footer } from "./Footer";
-import { SideBar } from "./SideBar";
-import { Header } from "./Header";
 
 export const Layout = () => {
   return (
