+++ src/layout/routes/AdminRoutes.tsx
@@ -1,0 +1,5 @@
+import { Route, Routes } from "react-router-dom";
+
+export const AdminRoutes = () => {
+  return <div></div>;
+};
