+++ src/layout/routes/routes.tsx
@@ -1,0 +1,10 @@
+import { BrowserRouter as Router } from "react-router-dom";
+import { UserRoutes } from "./UserRoutes";
+
+export const AppRoutes = () => {
+  return (
+    <Router>
+      <UserRoutes />
+    </Router>
+  );
+};
